@@ -6,32 +6,32 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.18.2 <3.0.0"
+  sdk: ">=3.1.3 <4.0.0"
 
 dependencies:
-  cupertino_icons: ^1.0.5
-  dio: ^4.0.6
+  cupertino_icons: ^1.0.6
+  dio: ^5.3.3
   flutter:
     sdk: flutter
   flutter_iconly: ^1.0.2
-  google_fonts: ^3.0.1
+  google_fonts: ^6.1.0
   hive: ^2.2.3
   hive_flutter: ^1.1.0
-  intl: ^0.17.0
-  ionicons: ^0.2.1
+  intl: ^0.18.1
+  ionicons: ^0.2.2
   like_button: ^2.0.5
   pull_to_refresh: ^2.0.0
-  share_plus: ^6.1.0
-  url_launcher: ^6.1.6
+  share_plus: ^8.0.0
+  url_launcher: ^6.1.14
 
 dev_dependencies:
-  build_runner: ^2.3.2
-  flutter_launcher_icons: ^0.10.0
-  flutter_lints: ^2.0.1
+  build_runner: ^2.4.6
+  flutter_launcher_icons: ^0.13.1
+  flutter_lints: ^2.0.3
   flutter_test:
     sdk: flutter
-  hive_generator: ^2.0.0
-  rename: ^2.0.1
+  hive_generator: ^2.0.1
+  rename: ^3.0.0
 
 flutter_icons:
   ios: true
