@@ -1,5 +1,3 @@
 class AnimationSetting {
-
-  static Duration get animationFadeInOutDuration => const Duration(milliseconds: 500); 
-
+  static Duration get animationFadeInOutDuration => const Duration(milliseconds: 500);
 }
