@@ -1,29 +1,28 @@
 import 'package:flutter/material.dart';
 
 class SlideTopRoute extends PageRouteBuilder {
-
   final Widget? page;
 
   SlideTopRoute({this.page})
-  : super(
-      pageBuilder: (
-        BuildContext context,
-        Animation<double> animation,
-        Animation<double> secondaryAnimation,
-      ) =>
-          page!,
-      transitionsBuilder: (
-        BuildContext context,
-        Animation<double> animation,
-        Animation<double> secondaryAnimation,
-        Widget child,
-      ) =>
-          SlideTransition(
+      : super(
+          pageBuilder: (
+            BuildContext context,
+            Animation<double> animation,
+            Animation<double> secondaryAnimation,
+          ) =>
+              page!,
+          transitionsBuilder: (
+            BuildContext context,
+            Animation<double> animation,
+            Animation<double> secondaryAnimation,
+            Widget child,
+          ) =>
+              SlideTransition(
             position: Tween<Offset>(
               begin: const Offset(0, 1),
               end: Offset.zero,
             ).animate(CurvedAnimation(parent: animation, curve: Curves.fastOutSlowIn)),
             child: child,
           ),
-    );
+        );
 }
