@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 
 class SwipeBack {
-
-  void onHorizontalDragLeft( BuildContext context, DragUpdateDetails details) {
-    
+  void onHorizontalDragLeft(BuildContext context, DragUpdateDetails details) {
     const int sensitivity = 8;
 
     if (details.delta.dx > sensitivity) {
@@ -11,13 +9,11 @@ class SwipeBack {
     }
   }
 
-  void onVerticalDragDown( BuildContext context, DragUpdateDetails details) {
-    
+  void onVerticalDragDown(BuildContext context, DragUpdateDetails details) {
     const int sensitivity = 8;
 
     if (details.delta.dy > sensitivity) {
       Navigator.pop(context);
     }
   }
-  
 }
