@@ -6,13 +6,12 @@ import 'package:simple_news_app/models/news.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await Hive.initFlutter(); 
+  await Hive.initFlutter();
   Hive.registerAdapter(NewsAdapter());
   runApp(const NewsApp());
 }
 
 class NewsApp extends StatefulWidget {
-  
   const NewsApp({Key? key}) : super(key: key);
 
   @override
@@ -20,23 +19,13 @@ class NewsApp extends StatefulWidget {
 }
 
 class _NewsAppState extends State<NewsApp> {
-  
-  final ThemeData theme = ThemeData(
-    primaryColor: Colors.grey[200],
-    fontFamily: GoogleFonts.barlow().fontFamily
-  );
+  final ThemeData theme = ThemeData(primaryColor: Colors.grey[200], fontFamily: GoogleFonts.barlow().fontFamily);
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      theme: theme.copyWith(
-        splashColor: Colors.transparent,
-        colorScheme: theme.colorScheme.copyWith(
-          secondaryContainer: Colors.white,
-          secondary: Colors.black
-        )    
-      ),
+      theme: theme.copyWith(splashColor: Colors.transparent, colorScheme: theme.colorScheme.copyWith(secondaryContainer: Colors.white, secondary: Colors.black)),
       home: const App(),
     );
   }
