@@ -2,7 +2,6 @@ import 'package:hive/hive.dart';
 import 'package:simple_news_app/models/news.dart';
 
 class StorageService {
-
   Future<void> saveNewsInStorage(News news) async {
     Box box = await Hive.openBox<News>('news');
     box.add(news);
@@ -12,7 +11,7 @@ class StorageService {
     final Box<News> box = Hive.box<News>('news');
     final List<News> boxNews = box.values.toList();
     for (final News newsInBox in boxNews) {
-      if(newsInBox.url == news.url){
+      if (newsInBox.url == news.url) {
         box.delete(newsInBox.key);
       }
     }
@@ -22,5 +21,4 @@ class StorageService {
     Box<News> box = await Hive.openBox<News>('news');
     return box.values.toList();
   }
-
 }
